Move common interfaces into its own file
